Overhaul CPU layout to put ARM/Thumb implementation in a separate place.
This commit also adds a variety of other changes such as implementing
quite a few bitmap modes, text modes, and the beginnings for an affine
tile mode. No sprite functionality has been implemented, yet.

This commit also adds a the ability to give input to the emulator via
the keyboard.