Store large allocations on the heap to allow execution on windows.
By default, Windows has an extremely small stack size. This means that
large allocations on the stack will blow the stack, causing the program
to crash. This commit ensures all large allocations go on the heap,
which should allow Windows to properly run this emulator. These
allocations are not guaranteed to immedietly go on the heap, so this
emulator may still sometimes fail to run on Windows, depending on
compiler and compiler tuning knobs.

In addition, this commit contains a variety of optimizations, improving
performance and code quality.