Defer alignment responsibility to memory bus (memory controller).
This commit also implements unaligned SRAM reads/writes properly (they
are not automatically aligned or rotated in the same way that all other
memory addresses are). In order to handle this, we must defer the
alignment logic to the bus.

In particular, SRAM reads from unaligned addresses are not force-aligned
and then rotated in the same way that other unaligned reads are. They
read from the actual address requested, mirror the byte as necessary
(BYTE_BYTE) for hword and (BYTE_BYTE_BYTE_BYTE for word), and then do
any rotations as necessary (only particularly noticable for halfword
reads).