Implement basic CH1 APU logic.
CH1 is the tone+sweep channel. I have done an extremely basic validation
test in a number of games (emerald, fire red, link to the past, minish
cap), though a significantly more thorough validation test is needed.

This commit has no validation at all for the length counter, and running
a game which makes use of this feature in CH1 will throw an exception
for now.